crypto-shredding/python/requirements.txt
#pycryptodome
pycrypto